\name{MXM-internal}
\alias{InternalSES}
\alias{InternalMMPC}
\alias{IdentifyEquivalence}
\alias{apply_ideq}
\alias{compare_p_values}
\alias{identifyTheEquivalent}
\alias{max_min_assoc}
\alias{min_assoc}
\alias{nchoosekm}
\alias{univariateScore}
\alias{cat.ci}
\alias{condi}
\alias{condi.perm}
\alias{InternalSES.temporal}
\alias{InternalMMPC.temporal}
\alias{IdentifyEquivalence.temporal}
\alias{apply_ideq.temporal}
\alias{identifyTheEquivalent.temporal}
\alias{max_min_assoc.temporal}
\alias{min_assoc.temporal}
\alias{univariateScore.temporal}
\alias{is.sepset}

\alias{proc_time-class}

\title{ Internal MXM Functions }
\description{
Internal functions of Package \pkg{MXM}
}
\section{Functions}{
\itemize{
\item InternalSES( ... )
\item InternalMMPC( ... )
\item IdentifyEquivalence( ... )
\item apply_ideq( ... )
\item compare_p_values( ... )
\item identifyTheEquivalent( ... )
\item max_min_assoc( ... )
\item min_assoc( ... )
\item nchoosekm( ... )
\item univariateScore( ... )
\item cat.ci( ... )
\item condi( ... )
\item condi.perm( ... )
\item InternalSES.temporal( ... )
\item InternalMMPC.temporal( ... )
\item IdentifyEquivalence.temporal( ... )
\item apply_ideq.temporal( ... )
\item identifyTheEquivalent.temporal( ... )
\item max_min_assoc.temporal( ... )
\item min_assoc.temporal( ... )
\item univariateScore.temporal( ... )
\item is.sepset( ... ) 

}
}
\details{
  These functions are only for internal usage of the MXM package - NOT to be called by the user.
  
  For faster computations in the internal SES and MMPC functions, install the suggested package "\bold{gRbase}".
  %% or we have been too lazy to document
}
\keyword{Internal MXM Functions}