\name{MXM-package}
\alias{MXM-package}
\docType{package}
\title{
This is an R package that currently implements feature selection methods for identifying minimal, statistically-equivalent and equally-predictive feature subsets.
}
\description{
The 'MXM' (Mens eX Machina, meaning 'Mind from the Machine' in Latin) package provides source code for the SES algorithm and for some appropriate statistical conditional independence tests (testIndFisher, testIndLogistic, gSquare and censIndLR are included). Read the package's help pages for more details.
}
\details{
\tabular{ll}{
Package: \tab MXM\cr
Type: \tab Package\cr
Version: \tab 0.4.3\cr
Date: \tab 2015-07-28\cr
License: \tab GPL-2\cr
}
}
\author{
Giorgos Athineou <athineou@ics.forth.gr>, Vincenzo Lagani <vlagani@ics.forth.gr>, Giorgos Borboudakis <borbudak@ics.forth.gr>, Ioannis Tsamardinos <tsamard@ics.forth.gr>
}
\section{Maintainers }{
Giorgos Athineou <athineou@ics.forth.gr>, Vincenzo Lagani <vlagani@ics.forth.gr>
}
\references{
I. Tsamardinos, V. Lagani and D. Pappas (2012) Discovering multiple, equivalent biomarker signatures. In proceedings of the 7th conference of the Hellenic Society for Computational Biology & Bioinformatics - HSCBB12.
}
\keyword{ SES }
\keyword{ Multiple Feature Signatures }
\keyword{ Feature Selection }
\keyword{ Variable Selection }
\seealso{
\code{\link{SES},\link{censIndLR},\link{testIndFisher}, \link{testIndLogistic}, \link{gSquare}, \link{censIndLR}}
}

